Make NuGetRestore a dynamic component
Fixes: #2231
